+++ service/pings.go
@@ -15,13 +15,15 @@ import (
 	cruxIPFS "github.com/dedis/student_19_cruxIPFS"
 )
 
+// ExecReqPings sends all its own pings distances to the node that requested it
 func (s *Service) ExecReqPings(env *network.Envelope) error {
 
 	// Parse message
 	req, ok := env.Msg.(*cruxIPFS.ReqPings)
 	if !ok {
 		log.Error(s.ServerIdentity(), "failed to cast to ReqPings")
-		return errors.New(s.ServerIdentity().String() + " failed to cast to ReqPings")
+		return errors.New(s.ServerIdentity().String() +
+			" failed to cast to ReqPings")
 	}
 
 	// wait for pings to be finished
@@ -33,17 +35,18 @@ func (s *Service) ExecReqPings(env *network.Envelope) error {
 	myName := s.Nodes.GetServerIdentityToName(s.ServerIdentity())
 	// build reply
 	for peerName, pingTime := range s.OwnPings {
-		//if peerName == myName {
-		//	reply += myName + " " + peerName + " " + "0.0"
-		//} else {
-		reply += myName + " " + peerName + " " + fmt.Sprintf("%f", pingTime) + "\n"
-		//}
+		reply += myName + " " + peerName + " " +
+			fmt.Sprintf("%f", pingTime) + "\n"
 	}
 
 	log.Lvl3("sending", reply)
 	requesterIdentity := s.Nodes.GetByName(req.SenderName).ServerIdentity
 
-	e := s.SendRaw(requesterIdentity, &cruxIPFS.ReplyPings{Pings: reply, SenderName: myName})
+	e := s.SendRaw(requesterIdentity, &cruxIPFS.ReplyPings{
+		Pings:      reply,
+		SenderName: myName,
+	})
+
 	if e != nil {
 		panic(e)
 	}
@@ -51,21 +54,21 @@ func (s *Service) ExecReqPings(env *network.Envelope) error {
 	return e
 }
 
+// ExecReplyPings handle replies of other nodes ping distances to get the full
+// array of distances between each nodes
 func (s *Service) ExecReplyPings(env *network.Envelope) error {
 
 	// Parse message
 	req, ok := env.Msg.(*cruxIPFS.ReplyPings)
 	if !ok {
 		log.Error(s.ServerIdentity(), "failed to cast to ReplyPings")
-		return errors.New(s.ServerIdentity().String() + " failed to cast to ReplyPings")
+		return errors.New(s.ServerIdentity().String() +
+			" failed to cast to ReplyPings")
 	}
 
 	// process ping output
-	//log.LLvl1("resp=", req.Pings)
 	s.PingMapMtx.Lock()
 	lines := strings.Split(req.Pings, "\n")
-
-	//s.printDistances(s.Nodes.GetServerIdentityToName(s.ServerIdentity()) + " before update")
 
 	for _, line := range lines {
 		if line != "" {
@@ -89,7 +92,6 @@ func (s *Service) ExecReplyPings(env *network.Envelope) error {
 			s.PingDistances[dst][src] += pingRes
 		}
 	}
-	//s.printDistances(s.Nodes.GetServerIdentityToName(s.ServerIdentity()) + " after update")
 
 	s.PingMapMtx.Unlock()
 
@@ -125,17 +127,18 @@ func (s *Service) getPings(readFromFile bool) {
 		}
 		s.PingMapMtx.Unlock()
 
-		log.LLvl3(s.Nodes.GetServerIdentityToName(s.ServerIdentity()), "finished ping own meas with len", len(s.OwnPings))
-
-		observerNode := s.Nodes.GetServerIdentityToName(s.ServerIdentity())
-
-		//s.printDistances("After ownping " + observerNode)
+		log.LLvl3(s.Nodes.GetServerIdentityToName(s.ServerIdentity()),
+			"finished ping own meas with len", len(s.OwnPings))
 
 		// ask for pings from others
 		for _, node := range s.Nodes.All {
-			if node.Name != s.Nodes.GetServerIdentityToName(s.ServerIdentity()) {
-				//log.LLvl1(s.ServerIdentity().String(), s.Nodes.GetServerIdentityToName(s.ServerIdentity()), "sends raw to", node.Name, node.ServerIdentity)
-				e := s.SendRaw(node.ServerIdentity, &cruxIPFS.ReqPings{SenderName: s.Nodes.GetServerIdentityToName(s.ServerIdentity())})
+			if node.Name !=
+				s.Nodes.GetServerIdentityToName(s.ServerIdentity()) {
+
+				e := s.SendRaw(node.ServerIdentity, &cruxIPFS.ReqPings{
+					SenderName: s.Nodes.GetServerIdentityToName(
+						s.ServerIdentity())})
+
 				if e != nil {
 					panic(e)
 				}
@@ -143,77 +146,65 @@ func (s *Service) getPings(readFromFile bool) {
 			}
 		}
 
-		// wit for ping replies from everyone but myself
+		// wait for ping replies from everyone but myself
 		for s.NrPingAnswers != len(s.Nodes.All)-1 {
-			if s.Name == "node_1" {
-				log.LLvl1(s.Nodes.GetServerIdentityToName(s.ServerIdentity()), "got pings", s.NrPingAnswers)
-			}
 			time.Sleep(5 * time.Second)
 		}
 
-		//s.printDistances("Array full " + observerNode)
-
-		// prints
-		pingDistStr := observerNode + " pingDistStr--------> "
-
 		// divide all pings by 2
-
 		for i := 0; i < len(s.Nodes.All); i++ {
-			name1 := "node_" + strconv.Itoa(i)
+			name1 := NodeName + strconv.Itoa(i)
 			for j := 0; j < len(s.Nodes.All); j++ {
-				name2 := "node_" + strconv.Itoa(j)
-				s.PingDistances[name1][name2] = s.PingDistances[name1][name2] / 2.0
+				name2 := NodeName + strconv.Itoa(j)
+				s.PingDistances[name1][name2] =
+					s.PingDistances[name1][name2] / 2.0
 			}
 		}
 
-		//s.printDistances("After adaptation " + observerNode)
-
 		for i := 0; i < len(s.Nodes.All); i++ {
-			name1 := "node_" + strconv.Itoa(i)
+			name1 := NodeName + strconv.Itoa(i)
 			for j := 0; j < len(s.Nodes.All); j++ {
-				name2 := "node_" + strconv.Itoa(j)
-				if s.PingDistances[name1][name2] != s.PingDistances[name2][name1] {
+				name2 := NodeName + strconv.Itoa(j)
+				if s.PingDistances[name1][name2] !=
+					s.PingDistances[name2][name1] {
+
 					log.Lvl1("Error: ping not symmetric")
 				}
 			}
 		}
 
-		for i := 0; i < len(s.Nodes.All); i++ {
-			name1 := "node_" + strconv.Itoa(i)
-			for j := 0; j < len(s.Nodes.All); j++ {
-				name2 := "node_" + strconv.Itoa(j)
-				pingDistStr += name1 + "-" + name2 + "=" + fmt.Sprintf("%f ", s.PingDistances[name1][name2])
-			}
-			pingDistStr += "\n"
-		}
-
-		//log.LLvl1(pingDistStr)
-
 		// check that there are enough pings
 		if len(s.PingDistances) < len(s.Nodes.All) {
-			log.Lvl1(s.Nodes.GetServerIdentityToName(s.ServerIdentity()), " too few pings 1")
+			log.Lvl1(s.Nodes.GetServerIdentityToName(s.ServerIdentity()),
+				" too few pings 1")
 		}
 		for _, m := range s.PingDistances {
 			if len(m) < len(s.Nodes.All) {
-				log.Lvl1(s.Nodes.GetServerIdentityToName(s.ServerIdentity()), " too few pings 2")
+				log.Lvl1(s.Nodes.GetServerIdentityToName(s.ServerIdentity()),
+					" too few pings 2")
 				log.LLvl1(m)
 			}
 		}
 
-		log.Lvl2(s.Nodes.GetServerIdentityToName(s.ServerIdentity()), "has all pings, starting tree gen")
+		log.Lvl3(s.Nodes.GetServerIdentityToName(s.ServerIdentity()),
+			"has all pings, starting tree gen")
 
 		// check TIV just once
 		// painful, but run floyd warshall and build the static routes
 		// ifrst, let's solve the k means mistery
 
-		if s.Nodes.GetServerIdentityToName(s.ServerIdentity()) == "node_0" {
+		if s.Nodes.GetServerIdentityToName(s.ServerIdentity()) == Node0 {
 			for n1, m := range s.PingDistances {
 				for n2, d := range m {
 					//bestDist
 					for k := 0; k < len(s.Nodes.All); k++ {
-						namek := "node_" + strconv.Itoa(k)
-						if d > s.PingDistances[n1][namek]+s.PingDistances[namek][n2] {
-							log.LLvl1("TIV!", n1, n2, "through", namek, "original:", s.PingDistances[n1][n2], ">", s.PingDistances[n1][namek], "+", s.PingDistances[namek][n2])
+						namek := NodeName + strconv.Itoa(k)
+						if d > s.PingDistances[n1][namek]+
+							s.PingDistances[namek][n2] {
+							log.LLvl1("TIV!", n1, n2, "through", namek,
+								"original:", s.PingDistances[n1][n2], ">",
+								s.PingDistances[n1][namek], "+",
+								s.PingDistances[namek][n2])
 							break
 						}
 					}
@@ -254,45 +245,46 @@ func (s *Service) getPings(readFromFile bool) {
 	}
 }
 
+// measureOwnPings compute ping distances to all nodes from self
 func (s *Service) measureOwnPings() {
 	myName := s.Nodes.GetServerIdentityToName(s.ServerIdentity())
 	for _, node := range s.Nodes.All {
 
 		if node.ServerIdentity.String() != s.ServerIdentity().String() {
 
-			log.Lvl2(myName, "meas ping to ", s.Nodes.GetServerIdentityToName(node.ServerIdentity))
+			log.Lvl2(myName, "meas ping to ",
+				s.Nodes.GetServerIdentityToName(node.ServerIdentity))
 
 			for {
 
 				peerName := node.Name
-				pingCmdStr := "ping -W 150 -q -c 3 -i 1 " + node.ServerIdentity.Address.Host() + " | tail -n 1"
+				pingCmdStr := "ping -W 150 -q -c 3 -i 1 " +
+					node.ServerIdentity.Address.Host() + " | tail -n 1"
 				pingCmd := exec.Command("sh", "-c", pingCmdStr)
 				pingOutput, err := pingCmd.Output()
 				if err != nil {
 					log.Fatal("couldn't measure ping")
 				}
 
-				if strings.Contains(string(pingOutput), "pipe") || len(strings.TrimSpace(string(pingOutput))) == 0 {
-					log.Lvl1(s.Nodes.GetServerIdentityToName(s.ServerIdentity()), "retrying for", peerName, node.ServerIdentity.Address.Host(), node.ServerIdentity.String())
+				if strings.Contains(string(pingOutput), "pipe") ||
+					len(strings.TrimSpace(string(pingOutput))) == 0 {
+
+					log.Lvl1(s.Nodes.GetServerIdentityToName(
+						s.ServerIdentity()), "retrying for", peerName,
+						node.ServerIdentity.Address.Host(),
+						node.ServerIdentity.String())
 					log.LLvl1("retry")
 					continue
 				}
 
-				processPingCmdStr := "echo " + string(pingOutput) + " | cut -d ' ' -f 4 | cut -d '/' -f 1-2 | tr '/' ' '"
+				processPingCmdStr := "echo " + string(pingOutput) +
+					" | cut -d ' ' -f 4 | cut -d '/' -f 1-2 | tr '/' ' '"
 				processPingCmd := exec.Command("sh", "-c", processPingCmdStr)
 				processPingOutput, _ := processPingCmd.Output()
-				if string(pingOutput) == "" {
-					//log.Lvl1("empty ping ", myName + " " + peerName)
-				} else {
-					//log.Lvl1("%%%%%%%%%%%%% ping ", myName + " " + peerName, "output ", string(pingOutput), "processed output ", string(processPingOutput))
-				}
-
-				//log.Lvl1("%%%%%%%%%%%%% ping ", s.Nodes.GetServerIdentityToName(s.ServerIdentity()) + " " + peerName, "output ", string(pingOutput), "processed output ", string(processPingOutput))
 
 				strPingOut := string(processPingOutput)
 
 				pingRes := strings.Split(strPingOut, "/")
-				//log.LLvl1("pingRes", pingRes)
 
 				avgPing, err := strconv.ParseFloat(pingRes[5], 64)
 				if err != nil {
@@ -308,6 +300,7 @@ func (s *Service) measureOwnPings() {
 	}
 }
 
+// printDistances print ping distances in an array
 func (s *Service) printDistances(str string) {
 	str += "\n       | "
 	for i := 0; i < len(s.Nodes.All); i++ {
@@ -326,12 +319,14 @@ func (s *Service) printDistances(str string) {
 	log.Lvl1(str)
 }
 
+// printPings print ping distances in a list
 func (s *Service) printPings() {
 	str := "\n"
 	for i := 0; i < len(s.Nodes.All); i++ {
 		name1 := NodeName + strconv.Itoa(i)
 		for j := 0; j < len(s.Nodes.All); j++ {
 			name2 := NodeName + strconv.Itoa(j)
+			// ping node_1 node_4 = 65.0345
 			str += fmt.Sprintf("ping %s %s = %f\n", name1, name2,
 				s.PingDistances[name1][name2])
 		}
@@ -339,12 +334,14 @@ func (s *Service) printPings() {
 	log.Lvl1(str)
 }
 
+// writePingsToFile write ping distances to a file
 func (s *Service) writePingsToFile() {
 	str := ""
 	for i := 0; i < len(s.Nodes.All); i++ {
 		name1 := NodeName + strconv.Itoa(i)
 		for j := 0; j < len(s.Nodes.All); j++ {
 			name2 := NodeName + strconv.Itoa(j)
+			// node_5 node_7 : 108.2233
 			str += fmt.Sprintf("%s %s : %f\n", name1, name2,
 				s.PingDistances[name1][name2])
 		}
